requirements.txt
firecrawl-py>=0.0.16
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
supabase>=2.0.0